Overall Client/Server improvements, and VLM finetuning code (#16)
* Improvements to config through YAML and now VideoJob actually does video

* Simplified fps config, standardized to client defining its fps, and server only has fallback. Also set fps to float type overall

* Client/server fixes to shutdown handlers and fixes to make it run

* Client modifications

* Client fixes hopefully

* logging additions to client

* Apply keepalive fix for results websocket

* Server fix attempts

* Adjustments to server and client fixing

* more fixes

* Simplifications to hopefully fix things fully

* quick fix to server app

* Add video loading

* add video playback and other fixes

* Removing video loading since not needed

* small fixes to the server part

* Frontend changes

* Latest changes

* Added support for SmolVLM

* Latest changes to run with CUDA

* Better logs, and attempt at shutdown fixes

* Add test videos for benchmarks

* Added benchmarking to test inference speeds

* got video processing for data pre processing

* Benchmarking of models for inference, and data preprocessing

* Beginning of training code

* Mini training setup done

* Base training script (WIP, not working) for HF PEFT finetune

* Working setup for a small finetune on a part of the dataset

* fixed data processing, and started finetunes

* Small refactoring for code style

+++ src/iris/client/capture/camera.py
@@ -12,12 +12,16 @@ class CameraCapture:
     """
 
     def __init__(
-        self, camera_index: int = 0, width: int = 640, height: int = 480, fps: int = 30
+        self,
+        camera_index: int = 0,
+        width: int = 640,
+        height: int = 480,
+        fps: float = 30.0,
     ):
         self.camera_index = camera_index
         self.width = width
         self.height = height
-        self.fps = fps
+        self.fps = float(fps)
         self.cap = None
 
     def start(self) -> bool:
@@ -33,7 +37,7 @@ class CameraCapture:
         # Set properties - pylint: disable=no-member
         self.cap.set(cv2.CAP_PROP_FRAME_WIDTH, self.width)
         self.cap.set(cv2.CAP_PROP_FRAME_HEIGHT, self.height)
-        self.cap.set(cv2.CAP_PROP_FPS, self.fps)
+        self.cap.set(cv2.CAP_PROP_FPS, int(self.fps))
         return True
 
     def stop(self) -> None:
@@ -95,7 +99,9 @@ class CameraCapture:
 
         # Crop and resize frame if it doesn't match configured dimensions
         if frame.shape[1] != self.width or frame.shape[0] != self.height:
-            frame = self._crop_to_aspect_ratio(frame)  # Crop first to preserve aspect ratio
+            frame = self._crop_to_aspect_ratio(
+                frame
+            )  # Crop first to preserve aspect ratio
             frame = cv2.resize(frame, (self.width, self.height))  # Then resize
 
         encode_param = [int(cv2.IMWRITE_JPEG_QUALITY), quality]  # pylint: disable=no-member
